refactor(rn): break the ClickHouse dash require cycle
Metro warned on every app start:

  Require cycle: dash/adapters/clickhouse.tsx
    -> dash/clickhouse/clickhouseStream.ts
    -> dash/adapters/clickhouse.tsx

The adapter pulled buildStatsTotals and CH_CONTROLS from the stream, while
the stream pulled normalize plus the LogItem/RawLogRow types back out of the
adapter. Require cycles resolve to uninitialized values depending on which
module is evaluated first, so this was a real hazard rather than just noise.

Extract the pure log-row layer into dash/clickhouse/logItem.ts: RawLogRow,
LogItem, MetaFact, normalize, parseMetadataFacts and the private
formatRelativeTime/META_HIDDEN_KEYS. None of it touches JSX, so it belongs
below both modules rather than inside a .tsx.

The graph is now a DAG: logItem is a leaf; clickhouseStream and the
ErrorDigest/NamespaceFocus/NamespaceGrid views import it directly; the
adapter imports from both and re-exports the moved symbols so the
dash/index.ts barrel keeps its existing public API.

Verified: kbve-react-native:typecheck green, rn:lint 0 errors.
